Essential files for running the game
For this game to run, the directory containing the files should have the following structure:
PARENT
    bluddGame.htm
    gameStyles.css
    scripts
        bludgame.js
    viking pics
        bloodred.jpg
        skyblue.jpg
        Bludd.png
    red
        red queen.png
        red king.png
        red Rook.png
        red Knight.png
        red Bishop.png
        red pawn.png
    blue
        as the red directory, except sub blue for red.
Artist credit to Mark Behm of Wizards of the Coast for the image on Bludd.png. The image is a cropped portion of the art for the card Grothama, All-Devouring, from the "Battlebond" set of Magic: the Gathering (copyright 2018). Please don't sue me, Wizards. Your art is amazing.
The images on the chess pieces, though, are totally my own. I made them in Powerpoint.