Add CLI Bludd chess (gameplay.py)
This is a precursor to a version (hopefully soon to be released in the wild!) of Bludd for Bludd as an online game on my own Django-based website.

For now, I built this CLI application to test that the gameplay works. I used colorama to make it pretty, and I'm quite pleased with the result. It's very sparsely documented, but you have the ability to save and load games based on your own board states. En passant and castling are implemented.
There's also an image of the gameplay.